pyproject.toml
[project]
name = "anki-with-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nltk>=3.9.2",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
